+++ libfabric.vcxproj
@@ -521,6 +521,7 @@
     <ClCompile Include="prov\tcp\src\tcpx_cq.c" />
     <ClCompile Include="prov\tcp\src\tcpx_domain.c" />
     <ClCompile Include="prov\tcp\src\tcpx_rma.c" />
+    <ClCompile Include="prov\tcp\src\tcpx_msg.c" />
     <ClCompile Include="prov\tcp\src\tcpx_ep.c" />
     <ClCompile Include="prov\tcp\src\tcpx_fabric.c" />
     <ClCompile Include="prov\tcp\src\tcpx_eq.c" />
