+++ prov/tcp/Makefile.include
@@ -5,6 +5,7 @@ _tcp_files = \
 	prov/tcp/src/tcpx_conn_mgr.c	\
 	prov/tcp/src/tcpx_domain.c	\
 	prov/tcp/src/tcpx_rma.c		\
+	prov/tcp/src/tcpx_msg.c		\
 	prov/tcp/src/tcpx_ep.c		\
 	prov/tcp/src/tcpx_shared_ctx.c	\
 	prov/tcp/src/tcpx_cq.c		\
